combine tests into a test file against Modal
